# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SecureTrack ERP API",
    "version": "1.0.0",
    "description": "Auto-generated inventory of Next.js route handlers. Auth: session cookie (Supabase) unless public/worker."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase auth session cookies"
      },
      "WorkerSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-job-secret",
        "description": "JOB_WORKER_SECRET or CRON_SECRET"
      }
    }
  },
  "paths": {
...
    "/api/notifications/send": {
      "post": {
        "summary": "POST /api/notifications/send",
        "tags": [
          "notifications",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/security/dual-control": {
      "get": {
        "summary": "GET /api/security/dual-control",
        "tags": [
          "security",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "summary": "POST /api/security/dual-control",
        "tags": [
          "security",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  }
}